# backend/openapi.yaml
...
openapi: 3.0.3
info:
  title: KosherEats API
  version: "1.0.0"
  description: >
    HTTP API for the KosherEats / GreenEats food-delivery platform. Serves the
    consumer, seller, courier, and admin apps. All money fields are integer
    cents unless noted. Most paths are under /api/v1; auth is a Bearer JWT
    obtained from the /api/v1/auth endpoints.
servers:
  - url: http://localhost:8080
    description: Local dev
  - url: https://koshereats-api.fly.dev
    description: Production (Fly.io)
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/restaurants:
    get:
      tags: [restaurants]
      summary: List restaurants in the caller's vertical
      security: []
      parameters:
        - { $ref: "#/components/parameters/VerticalQuery" }
        - in: query
          name: lat
          schema: { type: number, format: double }
        - in: query
          name: lng
          schema: { type: number, format: double }
      responses:
        "200":
          description: Restaurant list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Restaurant" }
...
  /api/v1/restaurants/{id}:
    get:
      tags: [restaurants]
      summary: Get a single restaurant
      security: []
      parameters:
        - { $ref: "#/components/parameters/PathId" }
      responses:
        "200":
          description: Restaurant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Restaurant" }
        "404": { $ref: "#/components/responses/Error" }
...
  /api/v1/seller/restaurants:
    get:
      tags: [seller]
      summary: List the seller's restaurants
      responses:
        "200":
          description: Restaurants
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Restaurant" }
    post:
      tags: [seller]
      summary: Create a restaurant
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Restaurant" }
      responses:
        "200":
          description: Created restaurant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Restaurant" }
  /api/v1/seller/restaurant:
    get:
      tags: [seller]
      summary: Get a seller's restaurant (defaults to first owned)
      parameters:
        - { $ref: "#/components/parameters/SellerRestaurantQuery" }
      responses:
        "200":
          description: Restaurant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Restaurant" }
    put:
      tags: [seller]
      summary: Update a seller's restaurant
      parameters:
        - { $ref: "#/components/parameters/SellerRestaurantQuery" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Restaurant" }
      responses:
        "200":
          description: Updated restaurant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Restaurant" }
...
  /api/v1/seller/payouts:
    get:
      tags: [seller]
      summary: Per-order payout ledger (newest first, keyset-paginated)
      description: >
        One line per order that reached its successful terminal state
        (delivered; completed for pickup). Pass next_cursor back as cursor for
        the next page; next_cursor is absent on the last page.
      parameters:
        - { $ref: "#/components/parameters/SellerRestaurantQuery" }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - in: query
          name: cursor
          schema: { type: string }
      responses:
        "200":
          description: Ledger page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RestaurantPayoutList" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
  /api/v1/admin/restaurants:
    get:
      tags: [admin]
      summary: List all restaurants
      responses:
        "200":
          description: Restaurants
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Restaurant" }
    post:
      tags: [admin]
      summary: Create a restaurant
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Restaurant" }
      responses:
        "200":
          description: Created restaurant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Restaurant" }
...
  /admin/restaurants:
    get:
      tags: [admin]
      summary: No-login admin restaurants dashboard (HTML; ?key= shared secret)
      security: []
      parameters:
        - in: query
          name: key
          schema: { type: string }
      responses:
        "200": { description: HTML page }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PathId:
      in: path
      name: id
      required: true
      schema: { type: string }
...
    VerticalQuery:
      in: query
      name: vertical
      description: Branded app scope; defaults to 'kosher'. Ignored when authenticated (JWT claim wins).
      schema: { type: string, enum: [kosher, vegan] }
    SellerRestaurantQuery:
      in: query
      name: restaurant_id
      description: Target restaurant; omitted falls back to the seller's first owned restaurant.
      schema: { type: string }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    KosherCertification:
      type: string
      enum: [OU, OK, Kof-K, Star-K, cRc, Badatz, Chof-K, other]
...
    Restaurant:
      type: object
      properties:
        id: { type: string }
        owner_id: { type: string }
        name: { type: string }
        description: { type: string }
        image_url: { type: string }
        cover_image_url: { type: string }
        logo_url: { type: string }
        phone: { type: string }
        email: { type: string }
        street: { type: string }
        city: { type: string }
        state: { type: string }
        zip_code: { type: string }
        lat: { type: number, format: double }
        lng: { type: number, format: double }
        kosher_certification: { $ref: "#/components/schemas/KosherCertification" }
        certifying_agency: { type: string }
        is_cholov_yisroel: { type: boolean }
        is_pas_yisroel: { type: boolean }
        is_glatt_kosher: { type: boolean }
        kosher_certificate_url: { type: string }
        cuisine_type:
          type: array
          items: { type: string }
        rating: { type: number, format: double }
        review_count: { type: integer }
        delivery_fee: { type: integer, description: cents }
        min_order: { type: integer, description: cents }
        est_delivery_min: { type: integer, description: minutes }
        est_delivery_max: { type: integer, description: minutes }
        is_open: { type: boolean }
        is_active: { type: boolean }
        approval_status: { type: string, enum: [pending, approved, rejected] }
        delivery_mode: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    PayoutFulfillment:
      type: string
      description: >
        Fee class. courier_delivery = external provider or KE courier (KE fee
        10% of food subtotal; KE absorbs processing). pickup = 5% + actual
        Stripe processing fee. self_delivery = restaurant's own driver (5% +
        actual processing; restaurant keeps its full own delivery fee + tip).
      enum: [courier_delivery, pickup, self_delivery]
    RestaurantPayoutLine:
      type: object
      description: >
        All amounts are cents. food_subtotal_cents is the item subtotal after
        the (restaurant-funded) deal discount. sales_tax_cents is passed through
        100%. delivery_fee_cents / tip_cents are the amounts CREDITED to the
        restaurant (non-zero only for self_delivery). ke_fee_cents is the
        percentage fee only; processing_fee_cents is recorded for every class
        but deducted only for pickup / self_delivery, and is 0 until the sweep
        has read the charge from Stripe (net_cents is provisional until then).
        net_cents = food + tax + delivery + tip − ke_fee − (processing for
        pickup/self_delivery), never negative. order_number is omitted (orders
        have no human-facing number).
      properties:
        id: { type: string }
        order_id: { type: string }
        completed_at: { type: string, format: date-time }
        fulfillment: { $ref: "#/components/schemas/PayoutFulfillment" }
        food_subtotal_cents: { type: integer }
        sales_tax_cents: { type: integer }
        delivery_fee_cents: { type: integer }
        tip_cents: { type: integer }
        ke_fee_cents: { type: integer }
        processing_fee_cents: { type: integer }
        net_cents: { type: integer }
        status:
          type: string
          enum: [awaiting_account, pending, paid, failed, reversed, void]
        transfer_id: { type: string, nullable: true }
        paid_at: { type: string, format: date-time, nullable: true }
      required: [id, order_id, completed_at, fulfillment, food_subtotal_cents, sales_tax_cents,
                 delivery_fee_cents, tip_cents, ke_fee_cents, processing_fee_cents, net_cents,
                 status, transfer_id, paid_at]
    RestaurantPayoutList:
      type: object
      properties:
        lines:
          type: array
          items: { $ref: "#/components/schemas/RestaurantPayoutLine" }
        next_cursor: { type: string, description: Absent on the last page }
      required: [lines]
...